+++ articles/bib2html.py
@@ -2,13 +2,14 @@
 # coding=utf-8
 
 from pybtex.database.input import bibtex
-from jinja2 import Template
 import dbio
-def convert(g_cfg):
+def read_bibtex(g_cfg):
+    bib_filename = g_cfg["bib_file"]
+    db_filename = g_cfg["db_file"]
     parser = bibtex.Parser()
-    bib_data = parser.parse_file(g_cfg["bib_file"])
+    bib_data = parser.parse_file(bib_filename)
     entries = []
-    art_db = dbio.articles_db(g_cfg["db_file"])
+    art_db = dbio.articles_db(db_filename)
     for key in bib_data.entries:
         try:
             persons = bib_data.entries[key].persons[u'author']
@@ -19,20 +20,36 @@ def convert(g_cfg):
         fields = bib_data.entries[key].fields
         if u'title' in fields:
             entry.update({ u"title" : fields[u'title'], })
+        else:
+            entry.update({ u"title" : "", })
         if u'journal' in fields:
             entry.update({ u"journal" : fields[u'journal'], })
+        else:
+            entry.update({ u"journal" : "", })
         if u'year' in fields:
             entry.update({ u"year" : fields[u'year'], })
+        else:
+            entry.update({ u"year" : "", })
         tags = art_db.article_tags(key)
         if tags:
             entry.update({ u"tags" : tags, })
+        else:
+            entry.update({ u"tags" : {}, })
         entries.append(entry)
-    
+    return entries
+
+from jinja2 import Template
+def convert(g_cfg):
+    db_filename = g_cfg["db_file"]
+    template_filename = g_cfg["template_file"]
+    entries = read_bibtex(g_cfg)
+    art_db = dbio.articles_db(db_filename)
+
     cfg = {}
     cfg.update({u'title' : u'Articles',})
     cfg.update({u'tags' : art_db.tags(),})
     cfg.update({u'entries' : entries,})
-    tmpl = Template(open(g_cfg["template_file"],'r').read())
+    tmpl = Template(open(template_filename,'r').read())
     html = tmpl.render(cfg)
     return html.encode("utf-8")